+++ crates/client/rpc/src/lib.rs
@@ -18,12 +18,12 @@ use errors::StarknetRpcApiError;
 use jsonrpsee::core::{async_trait, RpcResult};
 use jsonrpsee::types::error::CallError;
 use log::error;
-use mc_deoxys::l2::get_highest_block_hash_and_number;
-use mc_deoxys::utility::get_config;
 use mc_genesis_data_provider::GenesisProvider;
 pub use mc_rpc_core::utils::*;
 pub use mc_rpc_core::{Felt, StarknetReadRpcApiServer, StarknetTraceRpcApiServer, StarknetWriteRpcApiServer};
 use mc_storage::OverrideHandle;
+use mc_sync::l2::get_config;
+use mc_sync::utility::get_highest_block_hash_and_number;
 use mp_block::BlockStatus;
 use mp_contract::class::ContractClassWrapper;
 use mp_felt::{Felt252Wrapper, Felt252WrapperError};
@@ -38,14 +38,12 @@ use sc_client_api::backend::{Backend, StorageProvider};
 use sc_client_api::BlockBackend;
 use sc_network_sync::SyncingService;
 use sc_transaction_pool::{ChainApi, Pool};
-use sc_transaction_pool_api::error::{Error as PoolError, IntoPoolError};
-use sc_transaction_pool_api::{TransactionPool, TransactionSource};
+use sc_transaction_pool_api::TransactionPool;
 use sp_api::ProvideRuntimeApi;
 use sp_arithmetic::traits::UniqueSaturatedInto;
 use sp_blockchain::HeaderBackend;
 use sp_core::H256;
 use sp_runtime::traits::{Block as BlockT, Header as HeaderT};
-use sp_runtime::transaction_validity::InvalidTransaction;
 use sp_runtime::DispatchError;
 use starknet_api::api_core::ClassHash;
 use starknet_api::block::BlockHash;
@@ -70,15 +68,18 @@ use crate::utils::{
 };
 
 /// A Starknet RPC server for Madara
+#[allow(dead_code)]
 pub struct Starknet<A: ChainApi, B: BlockT, BE, G, C, P, H> {
     client: Arc<C>,
     backend: Arc<mc_db::Backend<B>>,
     overrides: Arc<OverrideHandle<B>>,
+    #[allow(dead_code)]
     pool: Arc<P>,
     #[allow(dead_code)]
     graph: Arc<Pool<A>>,
     sync_service: Arc<SyncingService<B>>,
     starting_block: <<B>::Header as HeaderT>::Number,
+    #[allow(dead_code)]
     genesis_provider: Arc<G>,
     _marker: PhantomData<(B, BE, H)>,
 }
@@ -230,6 +231,8 @@ where
         Ok(rpc_state_diff)
     }
 
+    #[allow(dead_code)]
+    #[allow(dead_code)]
     fn try_txn_hash_from_cache(
         &self,
         tx_index: usize,
@@ -261,9 +264,6 @@ where
         }
     }
 }
-
-/// Taken from https://github.com/paritytech/substrate/blob/master/client/rpc/src/author/mod.rs#L78
-const TX_SOURCE: TransactionSource = TransactionSource::External;
 
 #[async_trait]
 impl<A, B, BE, G, C, P, H> StarknetWriteRpcApiServer for Starknet<A, B, BE, G, C, P, H>
@@ -872,7 +872,7 @@ where
         let block_hash = starknet_block.header().hash::<H>();
 
         let actual_status = if starknet_block.header().block_number
-            <= mc_deoxys::l1::ETHEREUM_STATE_UPDATE.lock().unwrap().block_number
+            <= mc_sync::l1::ETHEREUM_STATE_UPDATE.lock().unwrap().block_number
         {
             BlockStatus::AcceptedOnL1.into()
         } else {
@@ -891,18 +891,12 @@ where
                 .map(FieldElement::from)
                 .collect()
         };
-        let block_status = match self.backend.messaging().last_synced_l1_block_with_event() {
-            Ok(l1_block) => {
-                if l1_block.block_number >= starknet_block.header().block_number {
-                    BlockStatus::AcceptedOnL1
-                } else {
-                    BlockStatus::AcceptedOnL2
-                }
-            }
-            Err(e) => {
-                error!("Failed to get last synced l1 block, error: {e}");
-                Err(StarknetRpcApiError::InternalServerError)?
-            }
+        let block_status: BlockStatus = if starknet_block.header().block_number
+            <= mc_sync::l1::ETHEREUM_STATE_UPDATE.lock().unwrap().block_number
+        {
+            BlockStatus::AcceptedOnL1
+        } else {
+            BlockStatus::AcceptedOnL2
         };
 
         let parent_blockhash = starknet_block.header().parent_block_hash;
@@ -1164,7 +1158,7 @@ where
         let starknet_version = starknet_block.header().protocol_version;
 
         let actual_status = if starknet_block.header().block_number
-            <= mc_deoxys::l1::ETHEREUM_STATE_UPDATE.lock().unwrap().block_number
+            <= mc_sync::l1::ETHEREUM_STATE_UPDATE.lock().unwrap().block_number
         {
             BlockStatus::AcceptedOnL1.into()
         } else {
@@ -1324,7 +1318,7 @@ where
                 error!("Failed to parse continuation token: {:?}", e);
                 StarknetRpcApiError::InvalidContinuationToken
             })?,
-            None => types::ContinuationToken { block_n: from_block.into(), event_n: 0 },
+            None => types::ContinuationToken { block_n: from_block, event_n: 0 },
         };
 
         // Verify that the requested range is valid
@@ -1444,7 +1438,7 @@ where
         let block = get_block_by_block_hash(self.client.as_ref(), substrate_block_hash)?;
         let block_header = block.header();
         let block_number = block_header.block_number;
-        let block_hash: Felt252Wrapper = block_header.hash::<H>().into();
+        let block_hash: Felt252Wrapper = block_header.hash::<H>();
         let starknet_version = block_header.protocol_version;
 
         let chain_id = self.chain_id()?.0;
@@ -1479,7 +1473,7 @@ where
         };
 
         let (tx_index, _) =
-            block_txs_hashes.into_iter().enumerate().find(|(_, hash)| hash == &transaction_hash.into()).unwrap().into();
+            block_txs_hashes.into_iter().enumerate().find(|(_, hash)| hash == &transaction_hash).unwrap();
 
         let transaction = block.transactions().get(tx_index).ok_or_else(|| {
             log::error!("Failed to retrieve transaction at index {tx_index} from block with hash {block_hash:?}");
@@ -1601,10 +1595,10 @@ where
 
         let actual_fee = execution_infos.actual_fee.0.into();
 
-        let actual_status = if block_number <= mc_deoxys::l1::ETHEREUM_STATE_UPDATE.lock().unwrap().block_number {
-            TransactionFinalityStatus::AcceptedOnL1.into()
+        let actual_status = if block_number <= mc_sync::l1::ETHEREUM_STATE_UPDATE.lock().unwrap().block_number {
+            TransactionFinalityStatus::AcceptedOnL1
         } else {
-            TransactionFinalityStatus::AcceptedOnL2.into()
+            TransactionFinalityStatus::AcceptedOnL2
         };
 
         let execution_result = match execution_infos.revert_error.clone() {
@@ -1640,7 +1634,7 @@ where
         // TODO: use actual execution ressources
         let receipt = match transaction {
             mp_transactions::Transaction::Declare(_) => TransactionReceipt::Declare(DeclareTransactionReceipt {
-                transaction_hash: transaction_hash.into(),
+                transaction_hash,
                 actual_fee,
                 finality_status: actual_status,
                 block_hash: block_hash.into(),
@@ -1703,43 +1697,6 @@ where
 
         Ok(MaybePendingTransactionReceipt::Receipt(receipt))
     }
-}
-
-async fn submit_extrinsic<P, B>(
-    pool: Arc<P>,
-    best_block_hash: <B as BlockT>::Hash,
-    extrinsic: <B as BlockT>::Extrinsic,
-) -> Result<<P as TransactionPool>::Hash, StarknetRpcApiError>
-where
-    P: TransactionPool<Block = B> + 'static,
-    B: BlockT,
-    <B as BlockT>::Extrinsic: Send + Sync + 'static,
-{
-    pool.submit_one(best_block_hash, TX_SOURCE, extrinsic).await.map_err(|e| {
-        error!("Failed to submit extrinsic: {:?}", e);
-        match e.into_pool_error() {
-            Ok(PoolError::InvalidTransaction(InvalidTransaction::BadProof)) => StarknetRpcApiError::ValidationFailure,
-            _ => StarknetRpcApiError::InternalServerError,
-        }
-    })
-}
-
-async fn convert_tx_to_extrinsic<C, B>(
-    client: Arc<C>,
-    best_block_hash: <B as BlockT>::Hash,
-    transaction: UserTransaction,
-) -> Result<<B as BlockT>::Extrinsic, StarknetRpcApiError>
-where
-    B: BlockT,
-    C: ProvideRuntimeApi<B>,
-    C::Api: StarknetRuntimeApi<B> + ConvertTransactionRuntimeApi<B>,
-{
-    let extrinsic = client.runtime_api().convert_transaction(best_block_hash, transaction).map_err(|e| {
-        error!("Failed to convert transaction: {:?}", e);
-        StarknetRpcApiError::InternalServerError
-    })?;
-
-    Ok(extrinsic)
 }
 
 fn convert_error<C, B, T>(