+++ crates/client/sync/src/l2.rs
@@ -42,6 +42,7 @@ pub struct L2StateUpdate {
 lazy_static! {
     /// Shared latest L2 state update verified on L2
     pub static ref STARKNET_STATE_UPDATE: Mutex<L2StateUpdate> = Mutex::new(L2StateUpdate {
+    pub static ref STARKNET_STATE_UPDATE: Mutex<L2StateUpdate> = Mutex::new(L2StateUpdate {
         block_number: u64::default(),
         global_root: StarkHash::default(),
         block_hash: StarkHash::default(),
@@ -104,7 +105,7 @@ impl BlockHashEquivalence {
         if block_number > 0 && let Some(block_hash_substrate) = block_hash_substrate {
             BlockHashEquivalence {
                 madara: block_hash_madara,
-                substrate: Some(H256::from_str(&block_hash_substrate).unwrap()),
+                substrate: Some(H256::from_str(block_hash_substrate).unwrap()),
             }
         } else {
             BlockHashEquivalence {
@@ -151,7 +152,7 @@ pub async fn sync<B: BlockT>(
                 let block = fetch_block(&client, block_sender, current_block_number);
                 let state_update = fetch_state_and_class_update(
                     &client,
-                    Arc::clone(&overrides),
+                    Arc::clone(overrides),
                     state_update_sender,
                     class_sender,
                     current_block_number,
@@ -165,7 +166,7 @@ pub async fn sync<B: BlockT>(
                 Ok(()),
                 fetch_state_and_class_update(
                     &client,
-                    Arc::clone(&overrides),
+                    Arc::clone(overrides),
                     state_update_sender,
                     class_sender,
                     current_block_number,
@@ -234,8 +235,8 @@ async fn fetch_state_and_class_update<B: BlockT>(
     rpc_port: u16,
     bonsai_dbs: BonsaiDbs<B>,
 ) -> Result<(), String> {
-    let state_update = fetch_state_update(&provider, block_number, bonsai_dbs).await?;
-    let class_update = fetch_class_update(&provider, &state_update, overrides, block_number, rpc_port).await?;
+    let state_update = fetch_state_update(provider, block_number, bonsai_dbs).await?;
+    let class_update = fetch_class_update(provider, &state_update, overrides, block_number, rpc_port).await?;
 
     // Now send state_update, which moves it. This will be received
     // by QueryBlockConsensusDataProvider in deoxys/crates/node/src/service.rs
@@ -375,7 +376,6 @@ fn aggregate_classes(state_update: &StateUpdate) -> Vec<&FieldElement> {
                 .iter()
                 .map(|DeclaredContract { class_hash, compiled_class_hash: _ }| class_hash),
         )
-        .chain(state_update.state_diff.old_declared_contracts.iter().map(|class_hash| class_hash))
         .unique()
         .collect()
 }
@@ -389,13 +389,10 @@ fn is_missing_class(
     block_hash_substrate: H256,
     class_hash: Felt252Wrapper,
 ) -> bool {
-    match overrides
+    overrides
         .for_schema_version(&StarknetStorageSchemaVersion::Undefined)
         .contract_class_by_class_hash(block_hash_substrate, ClassHash::from(class_hash))
-    {
-        Some(_) => false,
-        None => true,
-    }
+        .is_none()
 }
 
 /// Notifies the consensus engine that a new block should be created.